+++ tests/testthat/test-obj.R
@@ -47,6 +47,11 @@ test_that("rgl works", {
     l <- piece3d("die_face", x=1:8, suit=c(1:6, 1:2), rank=1:8, cfg=dice_d8())
     expect_true(length(l) == 8)
 
+    clear3d()
+    l <- piece3d("die_face", x=rep(1:5, 4), y=rep(1:4, each=5),
+                 rank=1:20, suit=rep(1:6, length.out=20), cfg=dice_d20())
+    expect_true(length(l) == 20)
+
     close3d()
 })